updated fk column data type to match the parent table.

+++ sql/ddl/create_question_response.sql
@@ -3,9 +3,9 @@ DROP TABLE IF EXISTS dev_catalog.slv_cdm_hrs.question_response;
 CREATE TABLE dev_catalog.slv_cdm_hrs.question_response (
     question_response_id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     question_response_description STRING,
-    respondent_id INT,
-    question_id INT,
-    wave_id INT,
+    respondent_id BIGINT,
+    question_id BIGINT,
+    wave_id BIGINT,
     start_date DATE,
     end_date DATE,
     active BOOLEAN,
